+++ inference.py
@@ -5,9 +5,8 @@ import numpy as np
 import torch
 
 from utmosv2.runner import run_inference
+from utmosv2.settings import configure_defaults, configure_inference_args
 from utmosv2.utils import (
-    configure_defaults,
-    configure_inference_args,
     get_dataloader,
     get_dataset,
     get_inference_data,
