+++ utmosv2/loss/__init__.py
@@ -1,3 +1,3 @@
-from utmosv2.loss.losses import CombinedLoss, PairwizeDiffLoss
+from utmosv2.loss._losses import CombinedLoss, PairwizeDiffLoss
 
 __all__ = ["PairwizeDiffLoss", "CombinedLoss"]
