+++ utmosv2/preprocess/__init__.py
@@ -1,3 +1,3 @@
-from utmosv2.preprocess.preprocess import add_sys_mean, preprocess, preprocess_test
+from utmosv2.preprocess._preprocess import add_sys_mean, preprocess, preprocess_test
 
 __all__ = ["preprocess", "preprocess_test", "add_sys_mean"]
