+++ utmosv2/runner/__init__.py
@@ -1,4 +1,9 @@
 from utmosv2.runner._inference import run_inference
-from utmosv2.runner._train import run_train
+from utmosv2.runner._train import run_train, train_1epoch, validate_1epoch
 
-__all__ = ["run_train", "run_inference"]
+__all__ = [
+    "run_train",
+    "train_1epoch",
+    "validate_1epoch",
+    "run_inference",
+]
