+++ utmosv2/settings/__init__.py
@@ -1,0 +1,7 @@
+from utmosv2.settings.config import (
+    configure_args,
+    configure_defaults,
+    configure_inference_args,
+)
+
+__all__ = ["configure_args", "configure_defaults", "configure_inference_args"]
