+++ utmosv2/transform/_xymasking.py
@@ -21,7 +21,7 @@ class XYMasking:
         self.fill_value = fill_value
         self.p = p
 
-    def __call__(self, img: torch.tensor) -> torch.tensor:
+    def __call__(self, img: torch.Tensor) -> torch.Tensor:
         if np.random.rand() < self.p:
             return img
         _, width, height = img.shape
