+++ utmosv2/utils/_pure/metrics.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+
+def print_metrics(metrics: dict[str, float]):
+    print(", ".join([f"{k}: {v:.4f}" for k, v in metrics.items()]))
